pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pcsx2-rpc"
version = "1.0.0"
description = "Production-quality Discord Rich Presence for PCSX2 (PlayStation 2 Emulator)"
authors = [{ name = "PCSX2 RPC Contributors" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pypresence>=4.3.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "loguru>=0.7.0",
    "pillow>=10.0.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "pytest-mock>=3.0"]

[project.scripts]
pcsx2-rpc = "main:cli_entry"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
